requirements.txt

Pillow == 8.2.0

PySide2 == 5.15.2
matplotlib == 3.4.1
numpy == 1.19.0
simple_pid == 1.0.0
tifffile == 2021.3.17
